refactor: batch ctl methods
