+++ api/monitor/save_resources_list.go
@@ -1,0 +1,9 @@
+package monitor
+
+// SaveResourcesList stores resource ids of current batch in Redis
+func (ctl BatchController) SaveResourcesList(batchID string, resourceIDs []string) error {
+	if err := ctl.rdb.SAdd("batch:"+batchID+":resources", resourceIDs).Err(); err != nil {
+		return err
+	}
+	return nil
+}
