+++ api/monitor/types.go
@@ -6,18 +6,10 @@ import (
 )
 
 type BatchController struct {
-	rdb        *redis.Client
-	admin *kafka.AdminClient
+	rdb   *redis.Client
+	kadmin *kafka.AdminClient
 }
 
-func NewBatchController(rdb *redis.Client, admin *kafka.AdminClient) BatchController {
-	return BatchController{rdb, admin}
-}
-
-func (ctl BatchController) Redis() *redis.Client {
-	return ctl.rdb
-}
-
-func (ctl BatchController) Kafka() *kafka.AdminClient {
-	return ctl.admin
+func NewBatchController(rdb *redis.Client, kadmin *kafka.AdminClient) BatchController {
+	return BatchController{rdb, kadmin}
 }
